app/requirements.txt
docker-py
paho-mqtt
pyyaml
